# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.CreateLeadRequest:
    properties:
      custom_fields:
        additionalProperties: true
        type: object
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - custom_fields
    - email
    - name
    - phone
    type: object
...
  v1.SendReplyRequest:
    properties:
      message:
        type: string
    required:
    - message
    type: object
...
  v1.UpdateLeadRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    required:
    - email
    - name
    - phone
    type: object
...
  v1.UpdateStatusReques:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@crm.com
    name: CRM Support
  description: API untuk sistem CRM (Customer Relationship Management)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CRM Project API
  version: "1.0"
paths:
...
  /leads:
    get:
      description: Ambil daftar leads dengan filter dan pagination
      parameters:
      - description: Halaman (default 1)
        in: query
        name: page
        type: integer
      - description: Jumlah per halaman (default 10)
        in: query
        name: limit
        type: integer
      - description: Kata kunci pencarian
        in: query
        name: search
        type: string
      - description: Filter status lead
        in: query
        name: status
        type: string
      - description: Filter dari tanggal (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: Filter sampai tanggal (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List leads
      tags:
      - leads
    post:
      consumes:
      - application/json
      description: Buat lead baru ke sistem
      parameters:
      - description: Data lead
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat lead baru
      tags:
      - leads
  /leads/{id}:
    delete:
      description: Soft delete lead berdasarkan ID (admin only)
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus lead
      tags:
      - leads
    get:
      description: Ambil detail lead berdasarkan ID
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail lead
      tags:
      - leads
    put:
      consumes:
      - application/json
      description: Update nama, email, dan nomor HP lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      - description: Data lead yang diupdate
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateLeadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update data lead
      tags:
      - leads
  /leads/{id}/reply:
    post:
      consumes:
      - application/json
      description: Kirim balasan WhatsApp ke lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      - description: Pesan balasan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.SendReplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Balas pesan WhatsApp
      tags:
      - leads
  /leads/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update status lead (NEW, CONTACTED, dll)
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      - description: Status baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateStatusReques'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update status lead
      tags:
      - leads
  /leads/{id}/timeline:
    get:
      description: Ambil semua aktivitas, deals, dan tasks dari sebuah lead
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Timeline aktivitas lead
      tags:
      - leads
  /leads/aging:
    get:
      description: Ambil leads yang tidak disentuh lebih dari X hari
      parameters:
      - description: Jumlah hari (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lead yang sudah lama tidak diupdate
      tags:
      - leads
  /leads/import:
    post:
      consumes:
      - multipart/form-data
      description: Upload file CSV untuk import data leads
      parameters:
      - description: File CSV
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Import leads dari CSV
      tags:
      - leads
  /leads/trash:
    get:
      description: Ambil daftar lead yang sudah di-soft delete
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lead yang dihapus
      tags:
      - leads
  /leads/trash/{id}/restore:
    post:
      description: Pulihkan lead yang sudah dihapus (admin only)
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Restore lead
      tags:
      - leads
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"